uma: Make it possible to disable the zone bucket cache
Use a new zone flag to indicate that full buckets are not to be cached
but rather must be released to the "slab." This will be used for cache
zones for VM pages. We don't want to let pages accumulate in the bucket
cache since they are not counted in v_free_count and because caching
will diminish the number of pages available for superpage reservations.