Batch page queue insertions.
Concatenation of batch queues onto the main queue is done in O(1) time,
unlike other implementations of this optimization that I'm aware of. It
does this by not adding any state to pages in the batch queues. Rather,
any time a thread wants to dequeue a page it must concatenate the
corresponding batch first, ensuring that the page is in the main queue.
But the lack of per-page state makes this operation fast, so the
resulting overhead is low.

The batch queue size is bounded by vm.batch_max, which defaults to 64.